Remove control and hazmat (squash this not done)
* moved status out and deleted control as no longer needed
* removed hazmat
